+++ CHANGELOG.md
@@ -7,14 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
-## [3.11.0] - 2024-10-08
+## [3.11.0] - 2024-10-09
 
 >**IMPORTANT**: When updating from a version prior to 3.7.2, 3.6.0, 3.4.0 or 3.0.0 please see the update notes for those versions first!
 
 ### Changed
 
-- Reworked skill queue. It does now look more like the in-game skill queue.
-- Current trained skill (if any) now shown in character overview
+- Reworked skill queue to make it more useful.
+- Current skill in training (if any) is now shown in character overview
 - Now shows explicitly whether a character's training queue is active
 
 ### Fixed